02/04/21/16h

+++ ECCE/ECCE/Models/CarrinhoModel.cs
@@ -15,8 +15,7 @@ namespace ECCE.Models
         public int Quantidade { get; set; }
         public string Cor { get; set; }
         public string Tamanho { get; set; }
-
-        public decimal Preco { get; set; }
+        public double Preco { get; set; }
         public string Foto { get; set; }
     }
 }
